split up images.py for consistency
